Create avggoodness-withkmeans.py
Observes the overall average goodness of all the clusters through iterations. The cluster with the least (R+M)/len(cluster) value is picked and its member track with least Merit is removed from the data set each iteration. Also, in order to determine the fall in the curve (a point where good tracks are also being removed), a running average (window=4) is calculated and the computation for each agent stops if the respective difference in running avg values between the 1st and the 6th(8th or 10th) is positive.